NF: TaskData in a separate file
The main reason is that CollectionTask is already 2K lines long, so reducing it by 10% is still a readability improvement
