+++ AnkiDroid/src/main/java/com/ichi2/anki/CardBrowser.java
@@ -103,7 +103,7 @@ import java.util.regex.Pattern;
 
 import timber.log.Timber;
 import static com.ichi2.async.CollectionTask.TASK_TYPE.*;
-import static com.ichi2.async.CollectionTask.TaskData;
+import com.ichi2.async.TaskData;
 
 public class CardBrowser extends NavigationDrawerActivity implements
         DeckDropDownAdapter.SubtitleListener {
