+++ AnkiDroid/src/main/java/com/ichi2/anki/DeckOptions.java
@@ -72,7 +72,7 @@ import java.util.TreeMap;
 
 import timber.log.Timber;
 import static com.ichi2.async.CollectionTask.TASK_TYPE.*;
-import static com.ichi2.async.CollectionTask.TaskData;
+import com.ichi2.async.TaskData;
 
 /**
  * Preferences for the current deck.
