+++ AnkiDroid/src/main/java/com/ichi2/anki/ModelFieldEditor.java
@@ -42,7 +42,7 @@ import com.ichi2.utils.JSONObject;
 
 import java.util.ArrayList;
 import static com.ichi2.async.CollectionTask.TASK_TYPE.*;
-import static com.ichi2.async.CollectionTask.TaskData;
+import com.ichi2.async.TaskData;
 
 public class ModelFieldEditor extends AnkiActivity {
