+++ AnkiDroid/src/main/java/com/ichi2/anki/TemporaryModel.java
@@ -36,7 +36,7 @@ import com.ichi2.compat.CompatHelper;
 import com.ichi2.libanki.Model;
 import com.ichi2.utils.JSONObject;
 import static com.ichi2.async.CollectionTask.TASK_TYPE.*;
-import static com.ichi2.async.CollectionTask.TaskData;
+import com.ichi2.async.TaskData;
 
 
 public class TemporaryModel {
