+++ AnkiDroid/src/main/java/com/ichi2/anki/UIUtils.java
@@ -16,7 +16,7 @@ import java.util.Calendar;
 
 import timber.log.Timber;
 import static com.ichi2.async.CollectionTask.TASK_TYPE.*;
-import static com.ichi2.async.CollectionTask.TaskData;
+import com.ichi2.async.TaskData;
 
 public class UIUtils {
