+++ AnkiDroid/src/main/java/com/ichi2/anki/dialogs/CustomStudyDialog.java
@@ -56,7 +56,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
 import static com.ichi2.async.CollectionTask.TASK_TYPE.*;
-import static com.ichi2.async.CollectionTask.TaskData;
+import com.ichi2.async.TaskData;
 
 
 public class CustomStudyDialog extends AnalyticsDialogFragment {
